fixed chat feature
